+++ cmd/influxd/launcher/cmd.go
@@ -147,6 +147,7 @@ type InfluxdOpts struct {
 
 	ProfilingDisabled bool
 	MetricsDisabled   bool
+	UIDisabled        bool
 
 	NatsPort            int
 	NatsMaxPayloadBytes int
@@ -196,6 +197,7 @@ func NewOpts(viper *viper.Viper) *InfluxdOpts {
 
 		ProfilingDisabled: false,
 		MetricsDisabled:   false,
+		UIDisabled:        false,
 
 		StoreType:   BoltStore,
 		SecretStore: BoltStore,
@@ -560,6 +562,13 @@ func (o *InfluxdOpts) BindCliOpts() []cli.Opt {
 			Flag:    "metrics-disabled",
 			Desc:    "Don't expose metrics over HTTP at /metrics",
 			Default: o.MetricsDisabled,
+		},
+		// UI Config
+		{
+			DestP:   &o.UIDisabled,
+			Flag:    "ui-disabled",
+			Default: o.UIDisabled,
+			Desc:    "Disable the InfluxDB UI",
 		},
 	}
 }